### 2025-08-04 #8729 @editorialbot
[PRE REVIEW]: estim8 - An FMI-compliant Python toolbox for bioprocess modeling and parameter estimation
**Submitting author:** @MicroPhen (<a href="http://orcid.org/0000-0001-9784-3626">Stephan Noack</a>)
**Repository:** https://github.com/JuBiotech/estim8
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.1
