### 2026-01-30 cadf0a7
Add matrix with Python versions up to Python 3.13 to CI
