+++ .github/workflows/test.yml
@@ -7,7 +7,11 @@ on:
 
 jobs:
   test:
+    name: test (${{matrix.python-version}})
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12", "3.13"]
 
     steps:
     - name: Checkout repository
@@ -16,7 +20,8 @@ jobs:
     - name: Set up Miniconda
       uses: conda-incubator/setup-miniconda@v3
       with:
-        python-version: '3.10'
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
         activate-environment: testim8
         environment-file: environment.yml
         auto-activate-base: false
