### 2026-01-30 26f27c4
Pin ubuntu:24.04 in CI
