+++ .github/workflows/test.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   test:
     name: test (${{matrix.python-version}})
-    runs-on: ubuntu-latest
+    runs-on: ubuntu:24.04
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12", "3.13"]
