### 2026-02-06 bd90342
Pin OM action version, add omsimulator package
