### 2026-02-09 ba19271
Pin ubuntu-22:04 for CI
