### 2026-02-09 e25d820
Pin setup-openmodelica v1.0.2
