+++ .github/workflows/test.yml
@@ -27,7 +27,7 @@ jobs:
         auto-activate-base: false
 
     - name: setup OpenModelica compiler
-      uses: OpenModelica/setup-openmodelica@v1.0
+      uses: OpenModelica/setup-openmodelica@v1.0.2
       with:
         version: '1.25.5'
         packages: |
