### 2026-02-09 283696e
use stable setup-openmodelica
