### 2026-02-10 040fb69
Fix job order to avoid conda messing with omc compilers
