+++ .github/workflows/test.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   test:
     name: test (${{matrix.python-version}})
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12", "3.13"]
@@ -16,15 +16,6 @@ jobs:
     steps:
     - name: Checkout repository
       uses: actions/checkout@v4
-
-    - name: Set up Miniconda
-      uses: conda-incubator/setup-miniconda@v3
-      with:
-        auto-update-conda: true
-        python-version: ${{ matrix.python-version }}
-        activate-environment: testim8
-        environment-file: environment.yml
-        auto-activate-base: false
 
     - name: setup OpenModelica compiler
       uses: OpenModelica/setup-openmodelica@v1.0.2
@@ -36,10 +27,24 @@ jobs:
         libraries: |
           'Modelica 4.0.0'
 
+    - uses: actions/setup-python@v6
+      with:
+        python-version: 'pypy3.13'
+    - run: pip install OMPython
+
     - name: compile FMU
       shell: bash -l {0}
       run: |
         python tests/test_data/model_compilation.py --modelname Growth
+
+    - name: Set up Miniconda
+      uses: conda-incubator/setup-miniconda@v3
+      with:
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
+        activate-environment: testim8
+        environment-file: environment.yml
+        auto-activate-base: false
 
     - name: Run tests
       shell: bash -l {0}
