### 2026-02-10 f2314ea
PIN setup-python pypy3.10
