+++ .github/workflows/test.yml
@@ -29,7 +29,7 @@ jobs:
 
     - uses: actions/setup-python@v6
       with:
-        python-version: 'pypy3.13'
+        python-version: 'pypy3.10'
     - run: pip install OMPython
 
     - name: compile FMU
