### 2026-02-10 b8d66a3
Drop omsimulator install
