### 2026-02-11 bef0129
Adjust paper structue to new requirements
