+++ paper/paper.md
@@ -40,32 +40,32 @@ bibliography: literature.bib
 ---
 
 # Summary
-Modeling and simulation are indispensable tools for understanding the complex nature of biological systems and making the most of the information contained in experimental data. Functional integration of quantitative measurement data and rigorous parameter estimation are essential for creating valid, predictive models. We present $\texttt{estim8}$, a Python package for FMI-compliant ODE and DAE (bio)process modeling and parameter estimation. It offers convenient handling of multiple experimental replicates coupled with highly scalable solutions using federated simulation operations. The source code is available on [GitHub](https://github.com/JuBiotech/estim8). The package is unit-tested on Windows and Linux. A thorough documentation is available under [https://estim8.readthedocs.io](https://estim8.readthedocs.io) including various example notebooks.
+Modeling and simulation are indispensable tools for understanding the complex nature of biological systems and making the most of the information contained in experimental data. Functional integration of quantitative measurement data and rigorous parameter estimation are essential for creating valid, predictive models. We present $\texttt{estim8}$, a Python package for FMI-compliant ODE and DAE (bio)process modeling and parameter estimation. It offers convenient handling of multiple experimental replicates coupled with highly scalable solutions using federated simulation operations. The source code is available on [GitHub](https://github.com/JuBiotech/estim8). The package is unit-tested on Windows and Linux. Thorough documentation is available under [https://estim8.readthedocs.io](https://estim8.readthedocs.io) including various example notebooks.
 
 # Statement of Need
-Mathematical modeling has become a pivotal tool in biotechnological research and industrial bioprocess development, supporting the analysis and interpretation of complex experimental data [@RN29;@RN27; @RN28;@RN38]. While ordinary differential equations (ODEs) are commonly used to describe continuous biological systems, many biotechnological applications require differential algebraic equation (DAE) systems to handle discontinuities, discrete events, physical constraints, and embedded optimization criteria [@RN26].
+Mathematical modeling has become a pivotal tool in biotechnological research and industrial bioprocess development, supporting the analysis and interpretation of complex experimental data [@RN29; @RN27; @RN28; @RN38]. While ordinary differential equations (ODEs) are commonly used to describe continuous biological systems, many biotechnological applications require differential algebraic equation (DAE) systems to handle discontinuities, discrete events, physical constraints, and embedded optimization criteria [@RN26].
 
-A crucial step in the modeling workflow is parameter estimation – or in layman terms "fitting the model". This step questions the theoretical understanding of the system under invesigation using real data, ultimately leading to confirmation or falsification of the hypotheses put forward. Although several general-purpose software tools for model formulation, simulation and parameter estimation exist, they currently have some limitations: Many only support ODE systems [@RN30; @RN33; @RN34; @RN31], require substantial workarounds for biological problems [@RN35; @RN36], or depend on proprietary licenses such as MATLAB for DAE simulation [@RN32].
+A crucial step in the modeling workflow is parameter estimation – or in layman's terms "fitting the model". This step questions the theoretical understanding of the system under investigation using real data, ultimately leading to confirmation or falsification of the hypotheses put forward. Although several general-purpose software tools for model formulation, simulation and parameter estimation exist, they currently have some limitations: Many only support ODE systems [@RN30; @RN33; @RN34; @RN31], require substantial workarounds for biological problems [@RN35; @RN36], or – although highly capable – pose a relatively high entry barrier [@RN32].
 
[…]
