### 2026-02-18 f44a0fe
Drop numpy pin, pin pandas<3
