+++ environment.yml
@@ -4,10 +4,10 @@ dependencies:
   - boa
   - conda-build
   - python>=3.10,<=3.13
-  - numpy>=1.17.0,<2
+  - numpy
   - pygmo==2.19.6
   - pip
-  - pandas
+  - pandas<3
   - openpyxl
   - matplotlib
   - seaborn
