+++ environment.yml
@@ -5,7 +5,7 @@ dependencies:
   - conda-build
   - python>=3.10,<=3.13
   - numpy
-  - pygmo==2.19.6
+  - pygmo>=2.19.6
   - pip
   - pandas<3
   - openpyxl
