### 2026-02-18 a7eda40
Drop boa and conda-build
