### 2026-02-18 #8729 @rkurchin
Thanks, @tobiaslatour, note that we still need a "State of the field" section, see more details in the docs [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain).
