### 2026-02-19 12173aa
Add pyPesto citation
