+++ paper/literature.bib
@@ -366,3 +366,23 @@
    volume = {9},
    year = {2023}
 }
+
+@article{RN48,
+    author = {Schälte, Yannik and Fröhlich, Fabian and Jost, Paul J and Vanhoefer, Jakob and Pathirana, Dilan and Stapor, Paul and Lakrisenko, Polina and Wang, Dantong and Raimúndez, Elba and Merkt, Simon and Schmiester, Leonard and Städter, Philipp and Grein, Stephan and Dudkin, Erika and Doresic, Domagoj and Weindl, Daniel and Hasenauer, Jan},
+    title = {pyPESTO: a modular and scalable tool for parameter estimation for dynamic models},
+    journal = {Bioinformatics},
+    volume = {39},
+    number = {11},
+    pages = {btad711},
+    year = {2023},
+    month = {11},
+    abstract = {Mechanistic models are important tools to describe and understand biological processes. However, they typically rely on unknown parameters, the estimation of which can be challenging for large and complex systems. pyPESTO is a modular framework for systematic parameter estimation, with scalable algorithms for optimization and uncertainty quantification. While tailored to ordinary differential equation problems, pyPESTO is broadly applicable to black-box parameter estimation problems. Besides own implementations, it provides a unified interface to various popular simulation and inference methods.pyPESTO is implemented in Python, open-source under a 3-Clause BSD license. Code and documentation are available on GitHub (https://github.com/icb-dcm/pypesto).},
+    issn = {1367-4811},
+    doi = {10.1093/bioinformatics/btad711},
+    url = {https://doi.org/10.1093/bioinformatics/btad711},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/39/11/btad711/53962204/btad711.pdf},
+}
+
+
+
+
