### 2026-02-19 19cfaed
Add State of the Field section
