+++ paper/literature.bib
@@ -382,7 +382,3 @@
     url = {https://doi.org/10.1093/bioinformatics/btad711},
     eprint = {https://academic.oup.com/bioinformatics/article-pdf/39/11/btad711/53962204/btad711.pdf},
 }
-
-
-
-
