### 2026-02-19 #8729 @tobiaslatour
Sorry @rkurchin , somehow i missed that. I just added a "State of the field" section.
