### 2026-02-20 #8729 @rkurchin
👋 @schmoelder, @blakeaw, and/or @ayush9pandey, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
