### 2026-02-21 #8729 @schmoelder
@rkurchin thanks for the invite, I'd love to review but since this is a package developed by my colleagues at IBG-1 (hi @microphen! ;-)), I'm too biased.
