### 2026-02-24 355c728
Update pyproject dependencies
