### 2026-02-24 d3e4e7c
move more deps to pip section in env yaml
