### 2026-02-25 9568239
bump version, pin pandas<3
