+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [
     ]
 description = "A parameter estimation tool for DAE models implemented compliant to the Functional Mockup Interface"
 readme = "README.md"
-requires-python = ">=3.10.0,<=3.13"
+requires-python = ">=3.10.0,<=3.14"
 license = {text = "AGPLv3"}
 classifiers=[
     "Programming Language :: Python :: 3",
