### 2026-02-27 2e3e0e9
Hotfix py13 support (#24)
* bump version

* Fix Python pin <3.14
