### 2026-03-02 32172d7
Bump OpenModelica/setup-openmodelica from 1.0.2 to 1.0.6
Bumps [OpenModelica/setup-openmodelica](https://github.com/openmodelica/setup-openmodelica) from 1.0.2 to 1.0.6.
- [Release notes](https://github.com/openmodelica/setup-openmodelica/releases)
- [Commits](https://github.com/openmodelica/setup-openmodelica/compare/v1.0.2...v1.0.6)

---
updated-dependencies:
- dependency-name: OpenModelica/setup-openmodelica
  dependency-version: 1.0.6
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
