+++ .github/workflows/test.yml
@@ -15,7 +15,7 @@ jobs:
 
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v4
+      uses: actions/checkout@v6
 
     - name: setup OpenModelica compiler
       uses: OpenModelica/setup-openmodelica@v1.0.2
