### 2026-03-04 #8729 @rkurchin
👋 @gchure, @mikeaalv, and/or @slabarthe, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
