### 2026-03-09 807b564
add ram threshold attibute and check during simulations
