### 2026-03-09 041bbe6
Add fmu ram threshold (#30)
* add ram threshold attibute and check during simulations

* bump version
