+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "estim8"
-version = "0.1.3"
+version = "0.1.4"
 authors = [
     {name = "Tobias Latour", email = "t.latour@fz-juelich.de"},
     {name = "Daniel Strohmeier", email = "d.strohmeier@fz-juelich.de"},
@@ -32,6 +32,7 @@ dependencies = [
     "pytensor",
     "grpclib",
     "openpyxl",
+    "psutil",
 ]
 
 [tool.setuptools.packages.find]
