+++ .github/workflows/test.yml
@@ -54,7 +54,7 @@ jobs:
         pytest -v --cov=estim8 --cov-report xml --cov-report term-missing estim8 tests/
 
     - name: Upload coverage
-      uses: codecov/codecov-action@v5
+      uses: codecov/codecov-action@v6
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
       with:
