### 2026-04-16 JuBiotech/estim8#33 @slabarthe
Error during example notebook execution
Hi,
After installation of estim8 following instructions [here](https://github.com/JuBiotech/estim8/blob/main/README.md#1-installation-from-pypi), I tried to execute your example notebook [1.modeling & simulation](https://github.com/JuBiotech/estim8/blob/main/docs/source/notebooks/1.%20Modeling%20%26%20Simulation.ipynb).

However, I got the following error:

While executing the first cell
```python
SimpleBatchModel = FmuModel(r'../../../tests/test_data/SimpleBatch.fmu')
```
I got the following error
```bash
Exception                                 Traceback (most recent call last)
Cell In[2], line 1
----> 1 SimpleBatchModel = FmuModel(r'../../../tests/test_data/SimpleBatch.fmu')

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/estim8/models.py:129, in FmuModel.__init__(self, path, fmi_type, default_parameters, r_tol, memory_threshold_mb)
    126 self.memory_threshold_mb = memory_threshold_mb
    128 # set fmi type and thereby load fmu
--> 129 self.fmi_type = fmi_type
    130 super().__init__(default_parameters=default_parameters, r_tol=r_tol)

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/estim8/models.py:226, in FmuModel.fmi_type(self, value)
    222     raise ValueError(
    223         f"{value} is not a valid FMI type. Choose one of 'CoSimulation' or 'Modelexchange'"
    224     )
    225 self._fmi_type = value
--> 226 self.instantiate_fmu()

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/estim8/models.py:184, in FmuModel.instantiate_fmu(self)
    182 self._model_description = fmpy.read_model_description(self.path)
    183 self._unzip_dir = fmpy.extract(self.path)
--> 184 self._fmu = fmpy.instantiate_fmu(
    185     self._unzip_dir, self._model_description, self._fmi_type
    186 )

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/fmpy/simulation.py:883, in instantiate_fmu(unzipdir, model_description, fmi_type, visible, debug_logging, logger, fmi_call_logger, library_path, early_return_allowed, event_mode_used, intermediate_update, require_functions, required_intermediate_variables)
    881     fmu.instantiate(functions=callbacks, loggingOn=debug_logging)
    882 elif is_fmi2:
--> 883     fmu = FMU2Model(**fmu_args)
    884     fmu.instantiate(visible=visible, callbacks=callbacks, loggingOn=debug_logging)
    885 else:

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/fmpy/fmi2.py:550, in FMU2Model.__init__(self, **kwargs)
    549 def __init__(self, **kwargs):
--> 550     super(FMU2Model, self).__init__(**kwargs)

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/fmpy/fmi2.py:103, in _FMU2.__init__(self, **kwargs)
    102 def __init__(self, **kwargs):
--> 103     super(_FMU2, self).__init__(**kwargs)

File ~/miniconda3/envs/estim8/lib/python3.10/site-packages/fmpy/fmi1.py:191, in _FMU.__init__(self, guid, modelIdentifier, unzipDirectory, instanceName, libraryPath, fmiCallLogger, requireFunctions)
    189 # check if shared library exists
    190 if not os.path.isfile(libraryPath):
--> 191     raise Exception("Cannot find shared library %s." % libraryPath)
    193 # change to the library directory as some DLLs expect this to resolve dependencies
    194 os.chdir(library_dir)

Exception: Cannot find shared library /tmp/tmpucwrmrgm/binaries/linux64/SimpleBatch.so.
```

Same kind of errors occured in the other notebooks while executing the FmuModel function.

I understand following this [issue](https://github.com/CATIA-Systems/FMPy/issues/330) that fmu models need some kind of compilation. I think that clear documentation for compiling it should be provided if it is up to the user to compile fmu models.

Could you provide any insight into why this is occurring?

Thank you in advance
