### 2026-04-16 #10147 @slabarthe
I could not execute properly the example notebooks. I opened an issue [here](https://github.com/JuBiotech/estim8/issues/33).
