### 2026-04-17 JuBiotech/estim8#33 @tobiaslatour
Hi @slabarthe

Thanks for opening the Issue.

FMUs need to be compiled for a target `os`, the one that is used in our examples is compiled for `win`.  I enriched the documentation on this: https://estim8.readthedocs.io/en/latest/fmi-export.html

For now, you could proceed by compiling the FMU on your System:

1. Install [OpenModelica](https://openmodelica.org/download/download-linux/)
2. Open `OMEdit`
3. [optional] If you want to run simulations in `CoSimulation` mode, change the FMI export setting: `Tools` -> `Options` -> `FMI` (left sidebar) -> set `Solver for Co-Simulation` to `CVODE`
4. Load the `SimpleBatch.mo` modelica model that is in the projects `test/test_data` directory  via  `File` -> `Open Model/Libraries Files(s)`
5. Export FMU:  Right-click on the loaded `SimpleBatch` class -> `Export` -> `FMU`
6. The console will log the filepath of the generated FMU, which you can then use

Please let me know if this solved your issue.
