### 2026-05-02 #10147 @slabarthe
> I could not execute properly the example notebooks. I opened an issue [here](https://github.com/JuBiotech/estim8/issues/33).

Detailed answer has been provided and dedicated page in documentation has been added.
