### 2026-05-05 d42513d
Fix tensor function creation for pytensor>=3.0.0
