### 2026-05-05 f8b735c
Fix typo in AI usage disclosure
