### 2026-05-06 7a6a50a
Add Python range to install instructions
