+++ tests/test_data/model_compilation.py
@@ -1,12 +1,14 @@
 import argparse
 
 
-def compileModel(modelname):
+def compileModel(modelname: str, output_name: str | None = None):
     modelpath = f"{os.getcwd()}/tests/test_data/{modelname}.mo"
     MO_model = ModelicaSystem(modelpath, modelname)
     # compile
     fmu = MO_model.convertMo2Fmu()
-    shutil.copy2(fmu, f"{os.getcwd()}/tests/test_data/{modelname}.fmu")
+    if output_name is None:
+        output_name = modelname
+    shutil.copy2(fmu, f"{os.getcwd()}/tests/test_data/{output_name}.fmu")
     return fmu
 
 
@@ -18,5 +20,6 @@ if __name__ == "__main__":
 
     parser = argparse.ArgumentParser()
     parser.add_argument("--modelname", required=True)
+    parser.add_argument("--output_name", default=None, required=False)
     args, _ = parser.parse_known_args()
-    compileModel(args.modelname)
+    compileModel(args.modelname, args.output_name)
