### 2026-05-06 3fe8098
Add detailed descriptions to figure legends
