### 2026-05-12 37d1066
Dont terminate FMU instance after simulation
