### 2026-07-10 cb7dc4d
Fix typo in conditions
