### 2026-07-10 1315405
Reformulate statement on UQ
