+++ paper/paper.md
@@ -63,7 +63,7 @@ The workflow in $\texttt{estim8}$ follows a structured approach to bioprocess mo
 
 The exported FMU is then loaded into $\texttt{estim8}$ using the $\texttt{FmuModel}$ class. The package implements a structured data hierarchy where $\texttt{Experiment}$ objects contain $\texttt{Measurement}$ objects with an associated $\texttt{error\_model}$ and an $\texttt{observation\_mapping}$. A key feature of $\texttt{estim8}$ is its comprehensive handling of biological replicates, which are crucial for the statistical quality of data from biological experiments [@RN37]. Based on a user-defined $\texttt{ParameterMapping}$, common properties of replicates and different conditions between them can be modeled by defining so-called global and local parameters. This concept utilizes redundant information in the measurement data, thereby effectively reducing the number of parameters to be estimated [@RN30; @RN47; @HelleckesOsthege2022].
 
-The core of $\texttt{estim8}$ is the $\texttt{Estimator}$ class, which serves as a central hub for managing parameter estimation tasks. This class stores all data entered by the user and provides functions for solving optimization problems. In addition, identifiability analyses and uncertainty quantifications can be performed using profile likelihood or Monte Carlo sampling. The $\texttt{visualization}$ module offers comprehensive visualization methods for analyzing simulation results, comparing model predictions with experimental data, and evaluating parameter estimation results. A thorough guideline is given by several example notebooks in our [documentation](https://estim8.readthedocs.io/en/latest/).
+The core of $\texttt{estim8}$ is the $\texttt{Estimator}$ class, which serves as a central hub for managing parameter estimation tasks. This class stores all data entered by the user and provides functions for solving optimization problems. In addition, identifiability analysis and uncertainty quantification can be performed using profile likelihood or Monte Carlo sampling. The $\texttt{visualization}$ module offers comprehensive visualization methods for analyzing simulation results, comparing model predictions with experimental data, and evaluating parameter estimation results. A thorough guideline is given by several example notebooks in our [documentation](https://estim8.readthedocs.io/en/latest/).
 
 ![](estim8_workflow.png)
 __Figure 1__: Schematic overview of the $\texttt{estim8}$ workflow.
