### 2026-07-10 6460976
Fix scipy citation title
