### 2026-07-10 7c71c84
Final paper corrections (#41)
* Remove unnecessary comma

* Fix typo in conditions

* Reformulate statement on UQ

* Fix scipy citation title
