### 2026-07-13 9bb4bb2
Bump version 0.1.6 (#42)
