### 2026-07-21 1b754ec
Make virtanen et al. sentence case
