### 2026-07-21 40e8380
Fix capitalization of Python and Modelica
