+++ paper/literature.bib
@@ -137,7 +137,7 @@
 
 @article{RN36,
    author = {K. Arendt, M. Jradi, M. Wetter, C.T. Veje},
-   title = {ModestPy: An Open-Source Python Tool for Parameter Estimation in Functional Mock-up Units},
+   title = {ModestPy: An Open-Source {Python} Tool for Parameter Estimation in Functional Mock-up Units},
    journal = {Proceedings of the American Modelica Conference 2018, Cambridge, MA, USA, October 9-10},
    doi = {10.3384/ecp18154121},
    year = {2018}
@@ -192,7 +192,7 @@
 
 @article{RN26,
    author = {Ploch, Tobias and Lieres, Eric von and Wiechert, Wolfgang and Mitsos, Alexander and Hannemann-Tamás, Ralf},
-   title = {Simulation of differential-algebraic equation systems with optimization criteria embedded in Modelica},
+   title = {Simulation of differential-algebraic equation systems with optimization criteria embedded in {Modelica}},
    journal = {Computers & Chemical Engineering},
    volume = {140},
    pages = {106920},
@@ -342,7 +342,7 @@
 
 @article{RN44,
     author = {Paul, Richard D and Jadebeck, Johann F and Stratmann, Anton and Wiechert, Wolfgang and Nöh, Katharina},
-    title = {hopsy — a methods marketplace for convex polytope sampling in Python},
+    title = {hopsy — a methods marketplace for convex polytope sampling in {Python}},
     journal = {Bioinformatics},
     volume = {40},
     number = {7},
