### 2026-07-21 9c15e43
Fix name format in Arendt et al.
