+++ paper/literature.bib
@@ -136,7 +136,7 @@
 }
 
 @article{RN36,
-   author = {K. Arendt, M. Jradi, M. Wetter, C.T. Veje},
+   author = {Arendt, K. and Jradi, M. and Wetter, M. and Veje, C.T.},
    title = {ModestPy: An Open-Source {Python} Tool for Parameter Estimation in Functional Mock-up Units},
    journal = {Proceedings of the American Modelica Conference 2018, Cambridge, MA, USA, October 9-10},
    doi = {10.3384/ecp18154121},
