### 2026-07-21 c423e7a
Fix a few citation formats
