+++ paper/literature.bib
@@ -34,7 +34,7 @@
 
 @article{RN35,
    author = { Nikolić, DD.},
-   title = {DAE Tools: equation-based object-oriented modelling, simulation and optimisation software.},
+   title = {{DAE Tools: equation-based object-oriented modelling, simulation and optimisation software.}},
    journal = {PeerJ Computer Science},
    volume = {2},
    number = {e54},
@@ -304,7 +304,7 @@
                Oldiges, Marco},
   journal   = {PLOS Computational Biology},
   publisher = {Public Library of Science},
-  title     = {Bayesian and calibration, process modeling and uncertainty quantification in biotechnology},
+  title     = {Bayesian calibration, process modeling and uncertainty quantification in biotechnology},
   year      = {2022},
   month     = {03},
   volume    = {18},
@@ -361,7 +361,7 @@
    doi = {10.7717/peerj-cs.1516},
    journal = {PeerJ Comput. Sci.},
    month = sep,
-   title = {{PyMC: a modern, and comprehensive probabilistic programming framework in Python}},
+   title = {{PyMC}: a modern, and comprehensive probabilistic programming framework in {Python}},
    url = {https://peerj.com/articles/cs-1516},
    volume = {9},
    year = {2023}
