+++ paper/literature.bib
@@ -137,7 +137,7 @@
 
 @article{RN36,
    author = {Arendt, K. and Jradi, M. and Wetter, M. and Veje, C.T.},
-   title = {ModestPy: An Open-Source {Python} Tool for Parameter Estimation in Functional Mock-up Units},
+   title = {ModestPy: An Open-Source {Python} Tool for Parameter Estimation in {Functional Mock-up Units}},
    journal = {Proceedings of the American Modelica Conference 2018, Cambridge, MA, USA, October 9-10},
    doi = {10.3384/ecp18154121},
    year = {2018}
@@ -281,7 +281,7 @@
 
 @misc{RN42,
    author = {{Modelica Association}},
-   title = {Functional Mock-up Interface (FMI)},
+   title = {{Functional Mock-up Interface (FMI)}},
    url = {https://fmi-standard.org/},
    type = {Web Page},
    year = {2023}
