### 2026-08-05 51b713c
Fix authors formatting of SMBL citation
