### 2026-08-05 ccbfa36
Joss corrections round 3 (#43)
* Make virtanen et al. sentence case

* Fix capitalization of Python and Modelica

* Fix name format in Arendt et al.

* Fix a few citation formats

* Capitalize Functional Mock-up Units/Interface

* Fix authors formatting of SMBL citation
