+++ paper/literature.bib
@@ -34,7 +34,7 @@
 
 @article{RN35,
    author = { Nikolić, DD.},
-   title = {DAE Tools: equation-based object-oriented modelling, simulation and optimisation software.},
+   title = {{DAE Tools: equation-based object-oriented modelling, simulation and optimisation software.}},
    journal = {PeerJ Computer Science},
    volume = {2},
    number = {e54},
@@ -136,8 +136,8 @@
 }
 
 @article{RN36,
-   author = {K. Arendt, M. Jradi, M. Wetter, C.T. Veje},
-   title = {ModestPy: An Open-Source Python Tool for Parameter Estimation in Functional Mock-up Units},
+   author = {Arendt, K. and Jradi, M. and Wetter, M. and Veje, C.T.},
+   title = {ModestPy: An Open-Source {Python} Tool for Parameter Estimation in {Functional Mock-up Units}},
    journal = {Proceedings of the American Modelica Conference 2018, Cambridge, MA, USA, October 9-10},
    doi = {10.3384/ecp18154121},
    year = {2018}
@@ -192,7 +192,7 @@
 
 @article{RN26,
    author = {Ploch, Tobias and Lieres, Eric von and Wiechert, Wolfgang and Mitsos, Alexander and Hannemann-Tamás, Ralf},
-   title = {Simulation of differential-algebraic equation systems with optimization criteria embedded in Modelica},
+   title = {Simulation of differential-algebraic equation systems with optimization criteria embedded in {Modelica}},
    journal = {Computers & Chemical Engineering},
    volume = {140},
    pages = {106920},
@@ -219,7 +219,7 @@
 
 @article{RN15,
    author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and van der Walt, Stéfan J. and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C. J. and Polat, İlhan and Feng, Yu and Moore, Eric W. and VanderPlas, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Antônio H. and Pedregosa, Fabian and van Mulbregt, Paul and Vijaykumar, Aditya and Bardelli, Alessandro Pietro and Rothberg, Alex and Hilboll, Andreas and Kloeckner, Andreas and Scopatz, Anthony and Lee, Antony and Rokem, Ariel and Woods, C. Nathan and Fulton, Chad and Masson, Charles and Häggström, Christian and Fitzgerald, Clark and Nicholson, David A. and Hagen, David R. and Pasechnik, Dmitrii V. and Olivetti, Emanuele and Martin, Eric and Wieser, Eric and Silva, Fabrice and Lenders, Felix and Wilhelm, Florian and Young, G. and Price, Gavin A. and Ingold, Gert-Ludwig and Allen, Gregory E. and Lee, Gregory R. and Audren, Hervé and Probst, Irvin and Dietrich, Jörg P. and Silterra, Jacob and Webber, James T. and Slavič, Janko and Nothman, Joel and Buchner, Johannes and Kulick, Johannes and Schönberger, Johannes L. and de Miranda Cardoso, José Vinícius and Reimer, Joscha and Harrington, Joseph and Rodríguez, Juan Lu
[…]
