### 2026-08-05 5bd8c70
Remove duplicate reference
