+++ CITATION.cff
@@ -1,0 +1,52 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Latour
+  given-names: Tobias
+  orcid: "https://orcid.org/0009-0009-7642-6853"
+- family-names: Strohmeier
+  given-names: Daniel
+  orcid: "https://orcid.org/0009-0008-9124-4153"
+- family-names: Osthege
+  given-names: Michael
+  orcid: "https://orcid.org/0000-0002-2734-7624"
+- family-names: Wiechert
+  given-names: Wolfgang
+  orcid: "https://orcid.org/0000-0001-8501-0694"
+- family-names: Noack
+  given-names: Stephan
+  orcid: "https://orcid.org/0000-0001-9784-3626"
+doi: 10.5281/zenodo.21337898
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Latour
+    given-names: Tobias
+    orcid: "https://orcid.org/0009-0009-7642-6853"
+  - family-names: Strohmeier
+    given-names: Daniel
+    orcid: "https://orcid.org/0009-0008-9124-4153"
+  - family-names: Osthege
+    given-names: Michael
+    orcid: "https://orcid.org/0000-0002-2734-7624"
+  - family-names: Wiechert
+    given-names: Wolfgang
+    orcid: "https://orcid.org/0000-0001-8501-0694"
+  - family-names: Noack
+    given-names: Stephan
+    orcid: "https://orcid.org/0000-0001-9784-3626"
+  date-published: 2026-08-06
+  doi: 10.21105/joss.10147
+  issn: 2475-9066
+  issue: 124
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 10147
+  title: estim8 - An FMI-compliant Python toolbox for bioprocess
+    modeling and parameter estimation
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.10147"
+  volume: 11
+title: estim8 - An FMI-compliant Python toolbox for bioprocess modeling
+  and parameter estimation
