### 2026-08-14 f3c0a1f
Add JOSS and zenodo badge (#44)
